changed chunk position to be a vector of ints

+++ game/src/Chunk.cpp
@@ -15,7 +15,7 @@
 #include "graphics/IndexBuffer.h"
 #include "constants.h"
 
-Chunk::Chunk(siv::PerlinNoise noise, glm::vec3 position, VertexArray *vao):
+Chunk::Chunk(siv::PerlinNoise noise, glm::ivec3 position, VertexArray *vao):
 	m_VertexArray(*vao),
 	m_TextureBuffer(GL_R8UI),
 	m_Noise(noise),
@@ -25,7 +25,7 @@ Chunk::Chunk(siv::PerlinNoise noise, glm::vec3 position, VertexArray *vao):
 {
 	LoadPosition(position);
 }
-Chunk::Chunk(siv::PerlinNoise noise, glm::vec3 position):
+Chunk::Chunk(siv::PerlinNoise noise, glm::ivec3 position):
 	m_VertexArray(),
 	m_TextureBuffer(GL_R8UI),
 	m_Noise(noise),
@@ -36,12 +36,12 @@ Chunk::Chunk(siv::PerlinNoise noise, glm::vec3 position):
 	LoadPosition(position);
 }
 
-void Chunk::LoadPosition(glm::vec3 position)
+void Chunk::LoadPosition(glm::ivec3 position)
 {
 	m_Position = position;
 	for (unsigned int x = 0; x < CHUNK_SIZE; x++) {
 		for (unsigned int z = 0; z < CHUNK_SIZE; z++) {
-			double h = m_Noise.noise3D_0_1((position.x * 16 + x) / 100.0, (position.z * 16 + z) / 100.0, 0.14342);
+			double h = m_Noise.noise3D_0_1((position.x * 16.0 + x) / 100.0, (position.z * 16.0 + z) / 100.0, 0.14342);
 			unsigned int height = static_cast<unsigned>(h * CHUNK_HEIGHT);
 			for (unsigned int y = 0; y < height; y++) {
 				m_ChunkData[x + y*CHUNK_SIZE + z * CHUNK_SIZE * CHUNK_HEIGHT] = 1;
@@ -116,13 +116,13 @@ void Chunk::CalculateMesh()
 		m_TextureBuffer.SetData(&m_VisibleFaces.front(), m_VisibleFaces.size() * sizeof(block_face), GL_STATIC_DRAW);
 }
 
-glm::vec3 Chunk::GetPositionChunkSpace()
+glm::ivec3 Chunk::GetPositionChunkSpace()
 {
 	return Chunk::m_Position;
 }
 
-glm::vec3 Chunk::GetCenterWorldSpace() {
-	return GetPositionWorldSpace() + glm::vec3(CHUNK_SIZE / 2, CHUNK_HEIGHT / 2, CHUNK_SIZE / 2);
+glm::ivec3 Chunk::GetCenterWorldSpace() {
+	return GetPositionWorldSpace() + glm::ivec3(CHUNK_SIZE / 2, CHUNK_HEIGHT / 2, CHUNK_SIZE / 2);
 }
 
 unsigned Chunk::GetVisibleFacesCount()
@@ -135,9 +135,9 @@ unsigned char Chunk::GetHeight()
 	return m_HeighestBlock;
 }
 
-glm::vec3 Chunk::GetPositionWorldSpace()
+glm::ivec3 Chunk::GetPositionWorldSpace()
 {
-	return glm::vec3(Chunk::m_Position.x * CHUNK_SIZE, 
+	return glm::ivec3(Chunk::m_Position.x * CHUNK_SIZE, 
 		Chunk::m_Position.y * CHUNK_HEIGHT, 
 		Chunk::m_Position.z * CHUNK_SIZE);
 }
